refactor: use UInt/ULong for OCI unsigned fields (K/N parity)
Match kontainer-runtime's Spec.kt: uid/gid → UInt, rlimit
hard/soft → ULong, id-mapping fields → UInt, seccomp arg
index → UInt, value/valueTwo → ULong, consoleSize → UInt.
Convert at FFM/syscall boundaries with toInt()/toLong().

+++ src/main/java/com/ternbusty/takoyaki/Main.kt
@@ -729,7 +729,7 @@ class Main {
                         System.err.println("takoyaki exec: bad console-size: $consoleSizeStr")
                         return 1
                     }
-                    consoleSize = ConsoleSize(width = w, height = h)
+                    consoleSize = ConsoleSize(width = w.toUInt(), height = h.toUInt())
                 }
             }
             return ExecCommand.run(
